+++ modules/graphics/functions/colormaps/cool.m
@@ -7,14 +7,18 @@
 % SPDX-License-Identifier: LGPL-3.0-or-later
 % LICENCE_BLOCK_END
 %=============================================================================
-function map = cool(m)
+function map = cool(varargin)
+  narginchk(0, 1);
   if nargin < 1
-    f = get(groot(), 'CurrentFigure');
+    R = groot();
+    f = R.CurrentFigure;
     if isempty(f)
-      m = 256;
+      m = size(R.DefaultFigureColormap, 1);
     else
-      m = size(f.Colormap,1);
+      m = size(f.Colormap, 1);
     end
+  else
+    m = varargin{1};
   end
   r = (0:m-1)' / max(m - 1, 1);
   map = [r, (1 - r), ones(m, 1)];