+++ modules/graphics/functions/colormaps/turbo.m
@@ -8,15 +8,15 @@
 % LICENCE_BLOCK_END
 %=============================================================================
 function varargout = turbo(varargin)
-  narginchk(0, 1);
   nargoutchk(0, 1);    
-  if nargin ==  0
-    gr  = groot();
-    f = gr.CurrentFigure; 
+  narginchk(0, 1);
+  if nargin < 1
+    R = groot();
+    f = R.CurrentFigure;
     if isempty(f)
-      n = 64;
+      n = size(R.DefaultFigureColormap, 1);
     else
-      n = size(f.Colormap,1);
+      n = size(f.Colormap, 1);
     end
   else
     n = varargin{1};